add null sink for discard-only MRT parsing
Add a --sink null option that parses MRT data and discards every entry,
useful for benchmarking the parser pipeline without a database backend.

- Add NullMRTEntriesWriteCloser (no-op MRTEntriesWriteCloser)
- Add NullIngestTask with progress/rate display configurers
- Wire SinkNull into the CLI alongside existing postgres and json-stdout
  sinks
- Change peer_as column from bigint to int in schema.sql

+++ schema.sql
@@ -33,7 +33,7 @@ CREATE TABLE IF NOT EXISTS mrt_entries (
     prefix_len  smallint NOT NULL,          -- net.IPNet.Mask.Size()
     peer        inet NOT NULL DEFAULT '::',       -- net.IP
     next_hop    inet NOT NULL DEFAULT '::',        -- net.IP
-    peer_as     bigint NOT NULL,            -- uint32，必须用 bigint 兼容 32-bit ASN
+    peer_as     int NOT NULL,            -- uint32，必须用 bigint 兼容 32-bit ASN
     as_path           int[] NOT NULL DEFAULT '{}',  -- postgres 的 int 对应 golang 的 int32，这里 golang 存的是 uint32
     community         int[] NOT NULL DEFAULT '{}',  -- []uint32，标准 BGP Community (RFC 1997)
     extended_community_high int[] NOT NULL DEFAULT '{}',  -- []uint32，BGP Extended Community (RFC 4360) 高 32 位
